Updated on 2026-09-14

+++ core/datasource/src/main/java/com/tangem/datasource/api/express/models/response/ExchangeProvider.kt
@@ -1,6 +1,7 @@
 package com.tangem.datasource.api.express.models.response
 
 import com.squareup.moshi.Json
+import java.math.BigDecimal
 
 data class ExchangeProvider(
     @Json(name = "id")
@@ -28,7 +29,7 @@ data class ExchangeProvider(
     val isRecommended: Boolean = false,
 
     @Json(name = "slippage")
-    val slippage: Int?,
+    val slippage: BigDecimal?,
 )
 
 enum class ExchangeProviderType {
